\name{example.of.Fisher.exact}
\alias{example.of.Fisher.exact}
\docType{data}
\title{Object of class `power.htest'}
\description{
Result from ss.nonadh(mu0=.1,mu1=.4,refinement="Fisher.exact"). It took about 8 seconds to calculate.
}
\usage{data(example.of.Fisher.exact)}
\details{
Compared with StatXact, gives same sample sizes.
}
%\source{
%NA
%}
%\references{
%NA
%}
%\examples{
%NA
%}
\keyword{datasets}
